styles: moved the employeedata to a new file and moved some of the styles
