+++ src/components/users.metadata.js
@@ -1,0 +1,67 @@
+const employeesData = [
+  {
+    id: 1,
+    name: "Joshua Gideon",
+    email: "joshgideon.com",
+    role: "Super Admin",
+    creationDate: "2017-01-02",
+    position: "CEO and Founder",
+    avatar: "images/img5.jpg",
+  },
+  {
+    id: 2,
+    name: "Henry Solomon",
+    email: "henrysolomon@gmail.com",
+    role: "Admin",
+    creationDate: "2020-11-04",
+    position: "Web developer",
+    avatar: "images/img3.png",
+  },
+  {
+    id: 3,
+    name: "Michael Orji",
+    email: "michaelorji.com",
+    role: "Admin",
+    creationDate: "2021-04-02",
+    position: "Graphics designer",
+    avatar: "images/img1.jpg",
+  },
+  {
+    id: 4,
+    name: "Gabriel Grace",
+    email: "gabrielgrace.com",
+    role: "HR Admin",
+    creationDate: "2009-09-04",
+    position: "Human Resource",
+    avatar: "images/img4.jpg",
+  },
+  {
+    id: 5,
+    name: "Tolulope Opeyemi",
+    email: "tolulopeopeyemi.com",
+    role: "Employee",
+    creationDate: "2022-03-14",
+    position: "Web designer",
+    avatar: "images/img3.png",
+  },
+  {
+    id: 6,
+    name: "Jeffery Benson",
+    email: "jeffbenson.com",
+    role: "Employee",
+    creationDate: "2011-06-20",
+    position: "Project Manager",
+    avatar: "images/img2.png",
+  },
+  {
+    id: 7,
+    name: "Habby Ogunlare",
+    email: "habbyogunlare.com",
+    role: "Admin",
+    creationDate: "2019-12-04",
+    position: "Web developer",
+    avatar: "images/img5.jpg",
+  },
+];
+
+export default employeesData;
