--- src/feature/Dashboard/Dashboard.module.css
@@ -1,14 +1,0 @@
-.container {
-  width: 100%;
-}
-
-.main {
-  padding: 76px 16px 16px 286px;
-  height: 100vh;
-  background: var(--bg-white);
-}
-
-.wrapper {
-  display: flex;
-  flex-direction: row;
-}
